refactor(desktop): keep channel status parsing outside component
