+++ desktop/src/renderer/src/features/settings/sections/ChannelsSection.tsx
@@ -2,37 +2,7 @@ import { useEffect, useState } from "react";
 import { useConnection } from "../../../stores/connection";
 import { Card, Empty, SectionHeader } from "./section-kit";
 import { StatusDot } from "../../../design/primitives";
-
-interface ChannelStatus {
-  name: string;
-  connected: boolean;
-}
-
-export function parseChannelStatusResponse(value: unknown): ChannelStatus[] {
-  const out: ChannelStatus[] = [];
-  if (Array.isArray(value)) {
-    for (const item of value) {
-      if (
-        item &&
-        typeof item === "object" &&
-        typeof (item as { name?: unknown }).name === "string"
-      ) {
-        out.push({
-          name: (item as { name: string }).name,
-          connected: Boolean((item as { connected?: unknown }).connected),
-        });
-      }
-    }
-    return out;
-  }
-  if (value && typeof value === "object") {
-    for (const [name, v] of Object.entries(value as Record<string, unknown>)) {
-      if (typeof v === "boolean") out.push({ name, connected: v });
-      else if (v && typeof v === "object") out.push({ name, connected: Boolean((v as { connected?: unknown }).connected) });
-    }
-  }
-  return out;
-}
+import { parseChannelStatusResponse, type ChannelStatus } from "./channel-status";
 
 export default function ChannelsSection() {
   const api = useConnection((s) => s.api);
